Allow unused expression values
The rationale [1] for this rule is that assigning intentionally unused
expressions to a discard improves code clarity and makes it easier to
spot unnecessary statements. However I find that using too many
explicitly discards tends to makes the code less readable overall so I
prefer not to enforce their use.

[1] https://learn.microsoft.com/en-us/dotnet/fundamentals/code-analysis/style-rules/ide0058#overview