Make database context available to GraphQL resolvers
As stated in the Hot Chocolate documentation [1], because Entity
Framework Core doesn't support multiple parallel operations being run on
the same context instance [2], access to the database context needs to
be synchronized. The simplest way to do this is to register the database
context using the `RegisterDbContext` method [3] from the
`HotChocolate.Data.EntityFramework` package.

[1] https://chillicream.com/docs/hotchocolate/v12/integrations/entity-framework
[2] https://learn.microsoft.com/en-gb/ef/core/miscellaneous/async
[3] https://chillicream.com/docs/hotchocolate/v12/integrations/entity-framework#registerdbcontext