Don't watch for config changes when running integration tests
By default [1] the ASP.NET `HostBuilder` configures the app to watch for
changes to the `appsettings.json` and `appsettings.{environment}.json`
config files. This can lead to the following error when running the
integration tests:

```
The configured user limit (128) on the number of inotify instances has
been reached, or the per-process limit on the number of open file
descriptors has been reached.
```

We can avoid this issue by setting `HostBuilder:ReloadConfigOnChange` to
`false` so that the config files are no longer watched by the app when
running the integration tests [2].

[1] https://github.com/dotnet/runtime/blob/97203d38ba734d76d39984c5206f2747f792fa89/src/libraries/Microsoft.Extensions.Hosting/src/HostingHostBuilderExtensions.cs#L230-L233
[2] https://learn.microsoft.com/en-us/aspnet/core/fundamentals/host/generic-host?view=aspnetcore-7.0#disable-app-configuration-reload-on-change