Add sample data factory with method for building a user
The internal counter is used to ensure that, as far as possible, every
property value generated by a single instance of the factory is unique.
The exceptions being null values and properties where the data type has
a small set of possible values.

Instance methods are not thread safe so a new instance of the factory is
expected to be created for each test.