Add utility methods for asserting on JSON elements
The `JsonAssert.Equivalent` methods are designed to reduce the amount of
boilerplate code needed to assert that the value of a given JSON
element, once deserialized, is equivalent to given object or collection
of objects of the same type. They can be used with anonymously typed
expected values, which will be particularly convenient for testing
GraphQL queries with projections and nesting.

`JsonAssert.Null`, on the other hand, provides only a small amount of
convenience over asserting on the `JsonElement.ValueKind` directly and
exists mainly for completeness.