Add fixture for tests that require a functional database context
This fixture can be used at class or collection level to provide a
database context factory that creates fully functional `AppDbContext`
instances connected to a test instance of the application database.

The type parameter `TDatabaseName` is used to generate the database
name. To ensure that only one test is run against each database instance
at a time, a unique `TDatabaseName` should be used for each collection
(or test class) that uses the fixture. A simple way to guarantee this is
to use the collection or test class itself.

This commit also introduces a `UseAppDbOptions` extension method on
`DbContextOptionsBuilder`. This is used to avoid duplicating the
`AppDbContext` configuration that would otherwise easily get out sync
between the application and the test fixture.